introduce a process-level lock file
Prevent multiple instances of buddha to run on the same machine at the same
time.

Since buddha is affecting the system services it's critical to understand it's
side-effects properly. The result of running multiple instances of buddha at
the same time is not well understood and might trigger some unexpected
race-conditions. The lock file prevents that.